pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "The server side for Somleng"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydub>=0.25.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.44",
    "supabase>=2.22.0",
    "uvicorn[standard]>=0.38.0",
]


[project.scripts]
backend = "app.main:main"
